+++ src/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
-import App from './app1'
+import App from './app'
 
 const container = document.getElementById('root');
 const root = createRoot(container!);
