+++ src/pages/GoodsPage.tsx
